+++ packages/url_launcher/url_launcher/lib/url_launcher.dart
@@ -44,7 +44,7 @@ import 'package:url_launcher_platform_interface/url_launcher_platform_interface.
 /// [enableDomStorage] is an Android only setting. If true, WebView enable
 /// DOM storage.
 /// [headers] is an Android only setting that adds headers to the WebView.
-/// [webOnlyLinkTarget] is an Web only setting . _blank opens the new url in new tab ,
+/// [webOnlyWindowName] is an Web only setting . _blank opens the new url in new tab ,
 /// _self opens the new url in current tab.
 /// Default behaviour is to open the url in new tab.
 ///
@@ -66,7 +66,7 @@ Future<bool> launch(
   bool universalLinksOnly,
   Map<String, String> headers,
   Brightness statusBarBrightness,
-  String webOnlyLinkTarget,
+  String webOnlyWindowName,
 }) async {
   assert(urlString != null);
   final Uri url = Uri.parse(urlString.trimLeft());
@@ -97,7 +97,7 @@ Future<bool> launch(
     enableDomStorage: enableDomStorage ?? false,
     universalLinksOnly: universalLinksOnly ?? false,
     headers: headers ?? <String, String>{},
-    webOnlyLinkTarget: webOnlyLinkTarget,
+    webOnlyWindowName: webOnlyWindowName,
   );
   assert(previousAutomaticSystemUiAdjustment != null);
   if (statusBarBrightness != null) {