+++ packages/url_launcher/url_launcher_platform_interface/lib/method_channel_url_launcher.dart
@@ -27,14 +27,16 @@ class MethodChannelUrlLauncher extends UrlLauncherPlatform {
   }
 
   @override
-  Future<bool> launch(String url,
-      {@required bool useSafariVC,
-      @required bool useWebView,
-      @required bool enableJavaScript,
-      @required bool enableDomStorage,
-      @required bool universalLinksOnly,
-      @required Map<String, String> headers,
-      String webOnlyLinkTarget}) {
+  Future<bool> launch(
+    String url, {
+    @required bool useSafariVC,
+    @required bool useWebView,
+    @required bool enableJavaScript,
+    @required bool enableDomStorage,
+    @required bool universalLinksOnly,
+    @required Map<String, String> headers,
+    String webOnlyWindowName,
+  }) {
     return _channel.invokeMethod<bool>(
       'launch',
       <String, Object>{