+++ packages/url_launcher/url_launcher_platform_interface/lib/url_launcher_platform_interface.dart
@@ -47,14 +47,16 @@ abstract class UrlLauncherPlatform extends PlatformInterface {
   ///
   /// For documentation on the other arguments, see the `launch` documentation
   /// in `package:url_launcher/url_launcher.dart`.
-  Future<bool> launch(String url,
-      {@required bool useSafariVC,
-      @required bool useWebView,
-      @required bool enableJavaScript,
-      @required bool enableDomStorage,
-      @required bool universalLinksOnly,
-      @required Map<String, String> headers,
-      String webOnlyLinkTarget}) {
+  Future<bool> launch(
+    String url, {
+    @required bool useSafariVC,
+    @required bool useWebView,
+    @required bool enableJavaScript,
+    @required bool enableDomStorage,
+    @required bool universalLinksOnly,
+    @required Map<String, String> headers,
+    String webOnlyWindowName,
+  }) {
     throw UnimplementedError('launch() has not been implemented.');
   }