+++ packages/url_launcher/url_launcher_web/lib/url_launcher_web.dart
@@ -34,25 +34,20 @@ class UrlLauncherPlugin extends UrlLauncherPlatform {
     UrlLauncherPlatform.instance = UrlLauncherPlugin();
   }
 
-  String _getUrlScheme(String url) =>
-      Uri
-          .tryParse(url)
-          ?.scheme;
+  String _getUrlScheme(String url) => Uri.tryParse(url)?.scheme;
 
   bool _isSafariTargetTopScheme(String url) =>
       _safariTargetTopSchemes.contains(_getUrlScheme(url));
 
-  /// Opens the given [url] in the specified [webOnlyLinkTarget].
+  /// Opens the given [url] in the specified [webOnlyWindowName].
   ///
   /// Returns the newly created window.
   @visibleForTesting
-  html.WindowBase openNewWindow(String url,{webOnlyLinkTarget} ) {
+  html.WindowBase openNewWindow(String url, {String webOnlyWindowName}) {
     // We need to open mailto, tel and sms urls on the _top window context on safari browsers.
     // See https://github.com/flutter/flutter/issues/51461 for reference.
-    final target =
-    browser.isSafari && _isSafariTargetTopScheme(url)
-        ? '_top'
-        : webOnlyLinkTarget ?? "";
+
+    final target = webOnlyWindowName ?? ((browser.isSafari && _isSafariTargetTopScheme(url)) ? '_top' : '');
     return _window.open(url, target);
   }
 
@@ -62,16 +57,16 @@ class UrlLauncherPlugin extends UrlLauncherPlatform {
   }
 
   @override
-  Future<bool> launch(String url, {
+  Future<bool> launch(
+    String url, {
     @required bool useSafariVC,
     @required bool useWebView,
     @required bool enableJavaScript,
     @required bool enableDomStorage,
     @required bool universalLinksOnly,
     @required Map<String, String> headers,
-    String webOnlyLinkTarget,
+    String webOnlyWindowName,
   }) {
-    return Future<bool>.value(
-        openNewWindow(url,webOnlyLinkTarget: webOnlyLinkTarget) != null);
+    return Future<bool>.value(openNewWindow(url, webOnlyWindowName: webOnlyWindowName) != null);
   }
 }