+++ packages/url_launcher/url_launcher_platform_interface/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.0.8
 
-* Added windowName web only parameter
+* Added webOnlyWindowName parameter
 
 ## 1.0.7